refactor: move helpers to common
